% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centerPlot.R
\name{centerPlot}
\alias{centerPlot}
\title{Creates center plot diplaying the cumulative number of centers that have been opened
as well as the cumulative number of centers that have been closed, if applicable, per trial week.}
\usage{
centerPlot(r)
}
\arguments{
\item{r}{recruitment scenario calculated with function \code{recruitment} of the interim package.}
}
\description{
Based on the function \code{trialCourse} and \code{eventCourse}.
}
\details{
Requires
}
