% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventPlot.R
\name{eventPlot}
\alias{eventPlot}
\title{Creates event plot diplaying the number of patients that have been screened and enrolled per trial week.
If the parameter \code{e1} are not \code{NULL}, then the number of events and number of drop-out before the first event is also displayed.}
\usage{
eventPlot(r, e1 = NULL)
}
\arguments{
\item{r}{recruitment scenario calculated with function \code{recruitment} of the interim package and}

\item{e1}{\emph{optional} for the simulation of the event simulation from function \code{event} of the interim package.}
}
\description{
Based on the function \code{eventCourse}.
}
\details{
Requires
}
