% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentPlot.R
\name{treatmentPlot}
\alias{treatmentPlot}
\title{Creates patient plot diplaying the number of patients that have been screened and enrolled per trial week.
If the parameter \code{t1} are not \code{NULL}, then the number of patients finished treatment is also displayed.}
\usage{
treatmentPlot(r, t1 = NULL)
}
\arguments{
\item{r}{recruitment scenario calculated with function \code{recruitment} of the interim package and}

\item{t1}{\emph{optional} for the simulation of the treatment phase simulation from function \code{treatment} of the interim package.}
}
\description{
Based on the function \code{trialCourse}.
}
\details{
Requires
}
